+++ tfjs-core/src/public/chained_ops/register_all_chained_ops.ts
@@ -18,6 +18,7 @@ import './add';
 import './all';
 import './any';
 import './arg_max';
+import './arg_min';
 import './atan2';
 import './avg_pool';
 import './batchnorm';
