+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -28,6 +28,7 @@ const CHAINED_OPS = [
   'all',
   'any',
   'argMax',
+  'argMin',
   'atan2',
   'avgPool',
   'batchNorm',
