+++ tfjs-core/src/register_all_gradients.ts
@@ -17,6 +17,7 @@
 import {addGradConfig} from './gradients/Add_grad';
 import {addNGradConfig} from './gradients/AddN_grad';
 import {argMaxGradConfig} from './gradients/ArgMax_grad';
+import {argMinGradConfig} from './gradients/ArgMin_grad';
 import {atan2GradConfig} from './gradients/Atan2_grad';
 import {avgPool3DGradConfig} from './gradients/AvgPool3D_grad';
 import {avgPoolGradConfig} from './gradients/AvgPool_grad';
@@ -72,6 +73,7 @@ const gradConfigs: GradConfig[] = [
   addGradConfig,
   addNGradConfig,
   argMaxGradConfig,
+  argMinGradConfig,
   atan2GradConfig,
   avgPoolGradConfig,
   avgPool3DGradConfig,
