+++ tfjs-core/src/tests.ts
@@ -44,6 +44,7 @@ import './ops/add_test';
 import './ops/all_test';
 import './ops/any_test';
 import './ops/arg_max_test';
+import './ops/arg_min_test';
 import './ops/arithmetic_test';
 import './ops/array_ops_test';
 import './ops/avg_pool_3d_test';
